Move endpoints to module
