+++ src/sentry/api/endpoints/codeowners/external_actor/team_index.py
