+++ src/sentry/api/endpoints/codeowners/external_actor/user_details.py
